Put scala pickling logic in the picklers package
* Help readability and mantainability of the code by separating the
  pickling logic from the nodes declaration.
